+++ client/src/components/LoginBadge.vue
@@ -1,14 +1,22 @@
 <script setup lang="ts">
-    import { getSession, login } from '@/model/session'
+    import { getSession, login, logout } from '@/model/session'
 
     const session =  getSession()
+
+    const doLogin = () => {
+        login('hbingley1@plala.or.jp', 'CQutx25i8r')
+    }
+
+    const doLogout = () => {
+        logout
+    }
 </script>
 
 <template>
     <div class="has-text-right" v-if="session.user">
         Welcome, {{ session.user.firstName }} {{ session.user.lastName }} <br>
         <small> {{ session.user.email }}
-            <a class="button is-small is-light is-warning" @click.prevent="session.user = null">
+            <a class="button is-small is-light is-warning" @click.prevent="doLogout">
                 <span class="icon">
                     <i class="fas fa-sign-out-alt"></i>
                 </span>
@@ -19,7 +27,7 @@
         <a class="button is-primary">
             <strong>Sign up</strong>
         </a>
-        <a class="button is-light" @click.prevent="login">
+        <a class="button is-light" @click.prevent="doLogin">
             Log in
         </a>
     </div>
